% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDistribution.R
\name{ClassDistribution}
\alias{ClassDistribution}
\title{A function to find out the data distribution from variables}
\usage{
ClassDistribution(
  inMat,
  Treatments,
  plots = TRUE,
  returnTable = FALSE,
  factorVector = NULL,
  PlotMain = NULL,
  ScalingFactorXaxis = NULL
)
}
\arguments{
\item{inMat}{a matrix with multiple treatments and variables from which the density shape and distribution of the variable values needs to be known.}

\item{Treatments}{a factor vector defining treatments in the input matrix, length must match the dimensions from the matrix.}

\item{plots}{defaults to TRUE Allows to return the diagnostic plots (ggplot dependency) that specify the distributions.}

\item{returnTable}{defaults to FALSE. Allows to return the final melted data frame used for density calculations.}

\item{factorVector}{defaults to NULL. a factor vector that defines the identity of independent factors in the design, in order to color them differently in the plots.}

\item{PlotMain}{defaults to NULL. When defined as a character vector it will appear as the diagnostic plots main title.}

\item{ScalingFactorXaxis}{defaults to NULL. numeric that is typically defined in the parent function as a factor to scale the x-axis in density plots with heavy-tailed distributions.}
}
\description{
This function allows to calculate density shapes and distributions from variable values in a multi-treatment matrix.
}
\examples{

...
}
\keyword{Distribution}
\keyword{ggplot}
\keyword{stats}
