% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMplotCustomizedMSI.R
\name{GLMplotCustomizedMSI}
\alias{GLMplotCustomizedMSI}
\title{A function to support generalized linear model based class comparison in replicated KineticMSI datasets}
\usage{
GLMplotCustomizedMSI(
  Variable,
  factorVector,
  Pvalues,
  MainTitle = "",
  ylabGLMs = NULL,
  xlabGLMs = NULL
)
}
\arguments{
\item{Variable}{Numeric vector where the means are inherited from to perform comparisons according to the factor vector.}

\item{factorVector}{factor type of vector that needs to define the treatments using the same nomenclature and order in the naming scheme as the names in the input variable.}

\item{Pvalues}{Vector of Q or P values that are used to return significance on top of the boxplots within the PDF containing the GLM results. Inherits from class comparison functions.}

\item{MainTitle}{Allows defining the title in the plots}

\item{ylabGLMs}{defaults to NULL. Character vector that defines the text on y-axes from GLM plots. Inherits from class comparison functions.}

\item{xlabGLMs}{defaults to NULL. Character vector that defines the text on x-axes from GLM plots. Inherits from class comparison functions.}
}
\description{
This function allows KineticMSI class comparison functions to create plots based on generalized linear models.
}
\examples{

...
}
\keyword{Class}
\keyword{Comparison}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Tracer}
