% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncorporationProxies.R
\name{IncorporationProxies}
\alias{IncorporationProxies}
\title{A function to produce files that describe through different proxies the tracer dynamics within kineticMSI datasets}
\usage{
IncorporationProxies(ParentDir, SteadyStatePoolsDir = NULL, ColSumNorm = FALSE)
}
\arguments{
\item{ParentDir}{directory where the input files are stored. This is also the address that will contain the isotope incorporation proxies as new csv files.}

\item{SteadyStatePoolsDir}{directory where the output steady state pool files are stored after running the function.}

\item{ColSumNorm}{defaults to FALSE. Allows to normalize the values of the calculated \emph{de novo} synthesized pools by the sum of all isotopologue pools, i.e., sum across feature of M0(A0) to Mn(An).}
}
\description{
This function allows to calculate across MSI pixels various values that reflect different aspects of the tracer dynamics. The function tests if the molecular features are shared across all datasets, if these are not shared, the function produces files with the common features before carrying on with the calculations. This is to prevent errors in the joined steady state pool files that are generated. The function outputs the isotope incorporation proxies as csv files to the same input directories.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{Incorporation}
\keyword{KineticMSI}
\keyword{Tracer}
