% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoCorTables.R
\name{IsoCorTables}
\alias{IsoCorTables}
\title{A function to produce the input files for Python-based correction of natural isotopic abundances (NIA)}
\usage{
IsoCorTables(PathToCSV)
}
\arguments{
\item{PathToCSV}{directory where the input file is stored.}
}
\description{
This function allows you to produce the input files needed for the Python tool \href{https://pypi.org/project/IsoCor/}{IsoCor} that corrects isotopologue envelopes for NIA. The function takes a single csv file with the right format and transforms it. Subsequently, the function generates a matrix to the R environment that can be exported in any desired format to serve as input for IsoCor.
}
\examples{

...
}
\keyword{IsoCor}
\keyword{KineticMSI}
\keyword{NIA}
\keyword{correction}
