% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kAssesmentMSI.R
\name{kAssesmentMSI}
\alias{kAssesmentMSI}
\title{A function to perform an initial step of replicated data quality assessment on KineticMSI datasets}
\usage{
kAssesmentMSI(
  path,
  PatternEnrichment = "MeanEnrichment",
  SubSetRepsIntensities = FALSE,
  CompareSampledSet = TRUE,
  returnObject = c(NULL, "RowMeansDataset", "minDataset"),
  factorVector,
  fun2clust = c("Enrichment", "AcqTime"),
  ZeroAction = c(NULL, "remove", "replace"),
  logiTransformation = FALSE,
  ScalingFactorXaxisDensityPlot = NULL
)
}
\arguments{
\item{path}{parent directory were all enrichment files are contained, digs within recursive folders.}

\item{PatternEnrichment}{defaults to "MeanEnrichment". Defines a character vector used to grab input csv enrichment files that can be later subset.}

\item{SubSetRepsIntensities}{defaults to FALSE. Allows to subset the MSI file list found in path.}

\item{CompareSampledSet}{defaults to TRUE. Compares the subsets and entire sets by plotting the ratios between mean and standard deviations across molecular features to allow evaluation of biased sampling. Optimally the ratios should oscillate around 1 with minimum variance.}

\item{returnObject}{can be either "RowMeansDataset" or "minDataset" and returns either a list with input matrices per molecular feature for the following mean class comparison (only the rowmeans per feature) or the minimum data subset to the minimum amount of pixels per entity.}

\item{factorVector}{character vector that needs to define the treatments using the same nomenclature and naming scheme as for the input files (follow the exemplary KineticMSI data for details).}

\item{fun2clust}{allows to decide whether pixels across files are sorted before subseting using their magnitude, i.e., "Enrichment" or their acquisition time, i.e., "AcqTime".}

\item{ZeroAction}{allows to define what happens with zeros within the entity data matrices. If null, the zeros are fully preserved, if "remove" the rows that only contain zeros are deleted and if "replace" all zeros are replaced by normally distributed randomly generated numbers in the scale of 10^-12 to 10^-13.}

\item{logiTransformation}{allows users to perform logit transformation before the class distribution assessment.}

\item{ScalingFactorXaxisDensityPlot}{defaults to NULL = 1 and defines the boundary of the x-axis in the ggplots that assess the distribution of the data.}
}
\description{
This function allows KineticMSI users to assess the quality in terms of reproducibility and mean distribution from measured molecular features across several independent treatments. The function first subsets all datasets to a common vector of molecular features and a common number of pixels. If molecular features are not the same across datatsets, the function produces new sets with the "Shared Features", which are named with this extension in the original IsoCorrectoR folders. Pixels are previously sorted per sample according to their magnitude, e.g., in the case of enrichment percentage from high to low enrichment. Afterwards a plot reflecting the ratio of means from subset and entire sets is produced in order to evaluate whether the subsetting procedure is skewing the data or whether the change trends will be fully preserved. An alternative to sorting by magnitude is sorting by acquisition time, which allows the users later on to evaluate differences that might be constrained in space or dependent on the instrument performance during a single sample. The function returns distribution plots in a PDF file across treatments that allow interpreting shifts in data that would otherwise remain unnoticed. Finally the function also returns either a list with the minimum datasets for all matrices or the compressed matrices used for mean distribution assessment and later for mean class comparison across samples.
}
\examples{

...
}
\keyword{Assessment}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Tracer}
