% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kClassComparisonMSI.R
\name{kClassComparisonMSI}
\alias{kClassComparisonMSI}
\title{A function that allows class comparison in replicated KineticMSI datasets}
\usage{
kClassComparisonMSI(
  kAssesmentOutput,
  factorVector,
  PDFname = "test",
  returnGLMplots = TRUE,
  patternGLMplot = c("Q values", "P values"),
  ylabTukey = NULL,
  xlabTukey = NULL,
  ylabGLM = NULL,
  xlabGLM = NULL
)
}
\arguments{
\item{kAssesmentOutput}{output from the previous KineticMSI function. Namely, kAssessmentMSI.R. The object must be a list of matrices, one matrix for each molecular feature of interest measured across replicates and treatments (i.e., rows in each matrix).}

\item{factorVector}{character vector that needs to define the treatments using the same nomenclature and order in the naming scheme as the row names in matrices within the input files (follow the exemplary KineticMSI data for details).}

\item{PDFname}{defaults to "test". Defines the name of the output PDF files that will be produce as outcome in the current working directory.}

\item{returnGLMplots}{defaults to TRUE. Defines whether GLM plots are returned.}

\item{patternGLMplot}{defaults to "Q values". Defines whether Q or P values are used to return significance on top of the boxplots within the PDF containing the GLM results.}

\item{ylabTukey}{defaults to NULL. Character vector that defines the text on y-axes from Tukey HSD plots.}

\item{xlabTukey}{defaults to NULL. Character vector that defines the text on x-axes from Tukey HSD plots.}

\item{ylabGLM}{defaults to NULL. Character vector that defines the text on y-axes from GLM plots.}

\item{xlabGLM}{defaults to NULL. Character vector that defines the text on x-axes from GLM plots.}
}
\description{
This function allows KineticMSI users to use classical class comparison algorithms such as ANOVA + Tukey HSD or generalized linear models (GLMs) + FDR correction. The input for the function comes from the previous kAssessmentMSI.R output. The function outputs graphics embedded in PDF files detailing the results from the class comparison algorithms. Additionally, a table with all the results is returned to the environment when the ouput is assigned to an object.
}
\examples{

...
}
\keyword{Class}
\keyword{Comparison}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Tracer}
