% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kClassDiscoveryMSI.R
\name{kClassDiscoveryMSI}
\alias{kClassDiscoveryMSI}
\title{A function to subset consolidated KineticMSI datasets into groups of related pixels}
\usage{
kClassDiscoveryMSI(
  path,
  PatternEnrichment = "MeanEnrichment",
  DistMethod = "abscor",
  nboot = 100,
  alpha = 0.9,
  SigClustHist = TRUE,
  SubSetRepsIntensities = FALSE,
  CompareSampledSet = TRUE,
  returnObject = c(NULL, "RowMeansDataset", "minDatasetPlusCoords"),
  fun2clust = c("Enrichment", "AcqTime"),
  ZeroAction = c(NULL, "remove", "replace"),
  logiTransformation = FALSE
)
}
\arguments{
\item{path}{parent directory were all enrichment files are contained, digs within recursive folders.}

\item{PatternEnrichment}{defaults to "MeanEnrichment". Defines a character vector used to grab input csv enrichment files that can be later subset.}

\item{DistMethod}{argument inherited from method.dist from pvclust.}

\item{nboot}{argument inherited from pvclust.}

\item{alpha}{defaults to 0.9, allows users to define the threshold AU-P value above which clusters are selected as significantly supported by the data. For details see pvclust package.}

\item{SigClustHist}{defaults to TRUE. Allows to return a histogram featuring the densities of significant cluster frequencies across molecular features.}

\item{SubSetRepsIntensities}{defaults to FALSE. Allows to subset the MSI file list found in path.}

\item{CompareSampledSet}{defaults to TRUE. Compares the subsets and entire sets by plotting the ratios between mean and standard deviations across molecular features to allow evaluation of biased sampling. Optimally the ratios should oscillate around 1 with minimum variance.}

\item{returnObject}{can be either "RowMeansDataset" or "minDatasetPlusCoords" and returns either a list with input matrices per molecular feature for the following mean class comparison (only the rowmeans per feature) or the minimum data subset to the minimum amount of pixels per entity.}

\item{fun2clust}{allows to decide whether pixels across files are sorted before subsetting using their magnitude, i.e., "Enrichment" or their acquisition time, i.e., "AcqTime".}

\item{ZeroAction}{allows to define what happens with zeros within the entity data matrices. If null, the zeros are fully preserved, if "remove" the rows that only contain zeros are deleted and if "replace" all zeros are replaced by normally distributed randomly generated numbers in the scale of 10^-12 to 10^-13.}

\item{logiTransformation}{allows users to perform logit transformation before the class distribution assessment.}
}
\description{
This function allows users to subset consolidated KineticMSI matrices (see our kAssesmentMSI.R function for details) into subsets that are validated internally by the data structure. The function clusters consolidated data matrices using a hierarchical clustering algorithm (HCA) with bootstrapping, which is a dependency from the R package pvclust. Then using a user defined significance threshold the function grabs an optimized number of significant clusters that feature AU-P values above the threshold. The optimization increases the threshold when many significant clusters are found until it optimizes the outcome to the minimum cluster number, or if there are any significant clusters then the optimization parameter lowers the threshold until obtaining significant partitions. The outcome from this tuning process is returned to the console. The function returns to the R environment a list of abundances and coordinates matrices per molecular feature that can be used to both map the clusters onto the original MSI images using kReconstructMSI.R or perform class comparison using the clustered subsets using kSubSetClassComparisonMSI.R. Additionally the function returns to the working directory a number of diagnostic plots that allow users to better contextualize the partitions in their datasets.
}
\examples{

...
}
\keyword{Clustering}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Significant}
\keyword{Subsetting}
\keyword{Tracer}
