% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kEnrichmentProportionsMSI.R
\name{kEnrichmentProportionsMSI}
\alias{kEnrichmentProportionsMSI}
\title{A function to compare proportions within selected tracer enrichment proxies across pixels in replicated KineticMSI datasets}
\usage{
kEnrichmentProportionsMSI(
  path,
  PatternEnrichment = "MeanEnrichment",
  SubSetRepsIntensities = FALSE,
  factorVector,
  ProportionOperator = c("equal", "less", "greater"),
  ProportionLimit = 0,
  kmeans = 5,
  KmBoot = 10,
  ClustMethod = "average",
  returnProprotionsHeatmap = TRUE
)
}
\arguments{
\item{path}{parent directory were all enrichment files are contained, digs within recursive folders.}

\item{PatternEnrichment}{defaults to "MeanEnrichment". Defines a character vector used to grab input csv enrichment files that can be later subset.}

\item{SubSetRepsIntensities}{defaults to FALSE. Allows to subset the MSI file list found in path.}

\item{factorVector}{character vector that needs to define the treatments using the same nomenclature and naming scheme as for the input files (follow the exemplary KineticMSI data for details).}

\item{ProportionOperator}{allows to define whether the proportions to be compared are "equal", "less" or "greater" than a value defined in the function call parameter "ProportionLimit".}

\item{ProportionLimit}{defaults to 0. Allows to define the Threshold value against which all comparisons are made.}

\item{kmeans}{defaults to 5. Allows to define the number of clusters looked for in each feature dataset.}

\item{KmBoot}{defaults to 10. Inherits from ComplexHeatmap and defines the number of bootstrap iterations used to build the K-mean consensus.}

\item{ClustMethod}{defaults to "average". Inherits from ComplexHeatmap and supports all clustering methods described there.}

\item{returnProprotionsHeatmap}{defaults to TRUE. Allows users to decide whether a comparison HeatMap representation is returned as a PDF file named "ProportionsHeatmap.pdf".}
}
\description{
This function allows KineticMSI users to compare the proportion of pixels that fall within specific ranges of selected tracer enrichment proxies. The user must define a threshold from which proportions are calculated and comparisons made using the parameter "ProportionLimit". Additionally the function provides the parameter "ProportionOperator" to define whether the comparisons are drawn in pixels "equal" to, "less" or "greater" than the predefined limit. The function returns to the R environment a matrix containing the values of the proportion comparison across molecular features (including statistical test outcomes) and an optional PDF with the graphical HeatMap representation of that matrix.
}
\examples{

...
}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Proportion}
\keyword{Replicates}
\keyword{Tracer}
