% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kReconstructMSI.R
\name{kReconstructMSI}
\alias{kReconstructMSI}
\title{A function to produce graphical reconstructions from MSI images using isotope tracer dynamics}
\usage{
kReconstructMSI(
  Reconstruct = c("After", "Before"),
  kClustersMSI,
  EnrPath,
  MSIPath = system.file("extdata", package = "KineticMSI"),
  PatternEnrichment = "MeanEnrichment.csv",
  outpath = getwd(),
  as = c("MSImageSet", "MSImagingExperiment"),
  PositionClusterLegend = "bottomleft",
  clustMethod = "average",
  clustDistance = "euclidean",
  kmeans = 5,
  KmBoot = 10,
  RevAbscissas = FALSE,
  RevOrdinates = FALSE,
  FactorName,
  yLabName = "Enrichment (\%)",
  paletteSpatialPlots = c("viridis", "magma", "plasma", "inferno", "cividis", "mako",
    "rocket", "turbo"),
  ContrastPercentValue = 0.1,
  SubSetRepsIntensities = FALSE,
  SubSetRepsMSI = FALSE,
  returnObject = TRUE
)
}
\arguments{
\item{Reconstruct}{either "Before" or "After". Allows defining if the reconstruction algorithm runs before or after applying the kClassDiscoveryMSI function.}

\item{kClustersMSI}{only needs to be defined if the previous parameter is set to "After". the input is a kClassDiscoveryMSI output object produced with the parameter returnObject set to "minDatasetPlusCoords".}

\item{MSIPath}{parent directory were all MSI files are contained, digs within recursive folders.}

\item{PatternEnrichment}{defaults to "MeanEnrichment.csv". character vector used to grab input csv enrichment files}

\item{outpath}{defaults to getwd(). Defines where the output files are written.}

\item{as}{inherits from Cardinal, type of MSI experimental file.}

\item{PositionClusterLegend}{defaults to "bottomleft". Allows to define the position of the legend in the cluster plot.}

\item{clustMethod}{defaults to "average". Inherits from ComplexHeatmap and supports all clustering methods described there.}

\item{clustDistance}{defaults to "euclidean". Inherits from ComplexHeatmap and supports all clustering methods described there.}

\item{kmeans}{defaults to 5. Allows to define the number of clusters looked for in each feature dataset.}

\item{KmBoot}{defaults to 10. Inherits from ComplexHeatmap and defines the number of bootstrap iterations used to build the K-mean consensus.}

\item{RevAbscissas}{defaults to FALSE. Allows to reverse the x-axis from its default position.}

\item{RevOrdinates}{defaults to FALSE. Allows to reverse the y-axis from its default position.}

\item{FactorName}{Character vector used to define the names of the plots x-axis.}

\item{yLabName}{Character vector used to define the names of the plots x-axis.}

\item{paletteSpatialPlots}{defines a palette inherited from viridis to color the intensities plots. Can be any of viridis, magma, plasma, inferno, cividis, mako, rocket, turbo.}

\item{ContrastPercentValue}{defaults to 0.1. Allows to define the minimum value to build contrast in the intensity plots.}

\item{SubSetRepsIntensities}{defaults to FALSE. Allows to subset the MSI file list found in path.}

\item{SubSetRepsMSI}{defaults to FALSE. Allows to subset the csv file list found in path.}

\item{returnObject}{defaults to TRUE. Allows to return the list of matrices to the R enviroment. When FALSE, only the output files are produced}

\item{Enrpath}{parent directory were all enrichment files are contained, digs within recursive folders.}
}
\description{
This function allows to produce the first type of inference on the replicated datasets that KineticMSI offers. The function produces first boxplots that reflect the dispersion of the chosen isotope tracer proxy across MSI pixels and treatments. The boxplots are produced both for separate replicates and as treatment means. Afterwards the function reconstructs the MSI acquired images by extracting the coordinates from the .ibd .imzML input files and using them to build reconstructed images that reflect the isotope tracer dynamics within the samples specimen. The function has two different run possibilities, it may be run "before" or "after" the KineticMSI class discovery function. In the former case the function will build clusters during the spatial reconstruction of images using a K-mean clustering algorithm and attempting at clustering together pixels based on specific similarity or dissimilarity metrics. In the latter case the clusters are inherited from the kClassDiscoveryMSI function output when the parameter returnObject is set to "minDatasetPlusCoords".
}
\examples{

...
}
\keyword{Dynamics}
\keyword{Image}
\keyword{Isotope}
\keyword{Reconstruction}
\keyword{Tracer}
