% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSubSetClassComparisonMSI.R
\name{kSubSetClassComparisonMSI}
\alias{kSubSetClassComparisonMSI}
\title{A function that allows class comparison in replicated KineticMSI data subsets}
\usage{
kSubSetClassComparisonMSI(
  kDiscoveryFactor1,
  kDiscoveryFactor2,
  factor1,
  factor2,
  repNumber1,
  repNumber2,
  PDFname,
  returnGLMplots = T,
  patternGLMplot = c("Q values", "P values"),
  ylabTukey = NULL,
  xlabTukey = NULL,
  ylabGLM = NULL,
  xlabGLM = NULL
)
}
\arguments{
\item{kDiscoveryFactor1}{output from the previous KineticMSI function. Namely, kClassDiscoveryMSI.R. The object must be a list of matrices, one matrix for each molecular feature of interest measured across replicates in factor1 (i.e., rows in each matrix).}

\item{kDiscoveryFactor2}{output from the previous KineticMSI function. Namely, kClassDiscoveryMSI.R. The object must be a list of matrices, one matrix for each molecular feature of interest measured across replicates in factor2 (i.e., rows in each matrix).}

\item{factor1}{character vector that needs to define the treatments using the same nomenclature and order in the naming scheme as the row names in matrices within the input files from kDiscoveryFactor1 (follow the exemplary KineticMSI data for details).}

\item{factor2}{character vector that needs to define the treatments using the same nomenclature and order in the naming scheme as the row names in matrices within the input files from kDiscoveryFactor2 (follow the exemplary KineticMSI data for details).}

\item{repNumber1}{number of biological replicates in dataset 1}

\item{repNumber2}{number of biological replicates in dataset 2}

\item{PDFname}{defaults to "test". Defines the name of the output PDF files that will be produce as outcome in the current working directory.}

\item{returnGLMplots}{defaults to TRUE. Defines whether GLM plots are returned.}

\item{patternGLMplot}{defaults to "Q values". Defines whether Q or P values are used to return significance on top of the boxplots within the PDF containing the GLM results.}

\item{ylabTukey}{defaults to NULL. Character vector that defines the text on y-axes from Tukey HSD plots.}

\item{xlabTukey}{defaults to NULL. Character vector that defines the text on x-axes from Tukey HSD plots.}

\item{ylabGLM}{defaults to NULL. Character vector that defines the text on y-axes from GLM plots.}

\item{xlabGLM}{defaults to NULL. Character vector that defines the text on x-axes from GLM plots.}
}
\description{
This function allows KineticMSI users to use classical class comparison algorithms such as ANOVA + Tukey HSD or generalized linear models (GLMs) + FDR correction. The input for the function comes from the previous kClassDiscoveryMSI.R output. The general assumption of this step is that multi-factorial designs were split at the class discovery level to allow discovery of inner data structures that are factor specific, which are then herein compared. The function outputs graphics embedded in PDF files detailing the results from the class comparison algorithms. Additionally, a table with all the results is returned to the environment when the output is assigned to an object.
}
\examples{

...
}
\keyword{Class}
\keyword{Comparison}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Subsets}
\keyword{Tracer}
