% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSummaryMSI.R
\name{kSummaryMSI}
\alias{kSummaryMSI}
\title{A function that allows summarizing the statistical output from KineticMSI full workflow}
\usage{
kSummaryMSI(
  kComparisonOutput,
  Abscissa = c("Cohensd", "Log2FC"),
  Ordinate = c("KS_pAdjusted", "Factor1F2_GLM_P values", "Factor1F2_GLM_Q values",
    "Factor1F2_TukeyHSD_Padj"),
  SigBndrie = 0.01,
  xBndrie = 2,
  label = c("Rownames", "EntityName"),
  plotTitle = "test",
  factor1,
  factor2,
  returnPlotsPNGs = TRUE,
  AbscissasName = "Abscissas",
  OrdinatesName = c("Ordinates", expression(-Log[10] ~ italic(P)[adj] ~ ~value)),
  width = 44.45,
  height = 27.78,
  AxesIndexSize = 10,
  AxesTitleSize = 10,
  LegendFontSize = 10,
  LegendKeySize = 5,
  LegendTitleSize = 10,
  LabelSize = 5,
  ColorAllDots = FALSE,
  DotsSize = 5,
  LowerLimitVolcano = NULL,
  UpperLimitVolcano = NULL,
  VolcanoLegendPosition = c("right", "bottom", "none", "top", "left", "right"),
  FisherTermsDir = paste0(system.file("extdata", package = "KineticMSI"),
    "/PA_Terms.csv"),
  returnSigFisher = FALSE,
  returnQvaluesFisher = FALSE,
  h1Fisher = c("two.sided", "greater", "less"),
  ColorCategory,
  PathwayColorsDir = c(NULL, paste0(system.file("extdata", package = "KineticMSI"),
    "/PA_TermsColors.txt")),
  BarNumbersPlot = c("pSigEntInClassToTotEntInClass", "pEntInClassToTotEnt",
    "SigEntInCat"),
  FisherCatNumSize = 3.5,
  FisherCatAxisFontSize = 10,
  FisherAxesTitleSize = 10,
  FisherLegendFontSize = 10,
  FisherLegendTitleSize = 10,
  FisherLegendKeySize = 10,
  LowerLimitFisher = NULL,
  UpperLimitFisher = NULL,
  FisherLegendPosition = c("right", "bottom", "none", "top", "left", "right"),
  returnObject = c("VolcanoPlots", "FisherPlots", "matrices")
)
}
\arguments{
\item{kComparisonOutput}{output from "kClassComparisonMSI.R" or "kSubSetClassComparisonMSI.R"}

\item{Abscissa}{name of the input column in "kComparisonOutput" to be used as the volcano abscissa.}

\item{Ordinate}{name of the input column in "kComparisonOutput" to be used as the volcano ordinate.}

\item{SigBndrie}{significance numerical boundary below which data points are labelled and significance vertical lines drawn in the volcano.}

\item{xBndrie}{numerical boundary above which data points are labelled and significance horizontal lines drawn in the volcano.}

\item{label}{parameter that allows to use as data point labels either the molecular entity name "EntityName" or the row name from "kComparisonOutput" when set to "Rownames".}

\item{plotTitle}{character vector inherited as a title for the volcano plots.}

\item{factor1}{character that must coincide with factor 1 used at the class comparison functions.}

\item{factor2}{character that must coincide with factor 2 used at the class comparison functions.}

\item{returnPlotsPNGs}{defaults to TRUE, builds .png files for every plot.}

\item{AbscissasName}{character to name the abscissa axis}

\item{OrdinatesName}{character to name the ordinate axis}

\item{width}{defaults to 44.45. graphical parameter for the saved .png volcano plots.}

\item{height}{defaults to 27.78. graphical parameter for the saved .png volcano plots.}

\item{AxesIndexSize}{defaults to 10 points.}

\item{AxesTitleSize}{defaults to 10 points.}

\item{LegendFontSize}{defaults to 10 points.}

\item{LegendKeySize}{defaults to 5 points.}

\item{LegendTitleSize}{defaults to 10 points.}

\item{LabelSize}{defaults to 5 points.}

\item{ColorAllDots}{defaults to FALSE. When TRUE colors all the plots with the category color, including those below the "SigBndrie" and "xBndrie".}

\item{DotsSize}{defaults to 5 points.}

\item{LowerLimitVolcano}{defaults to NULL. Allows to customize the lower limit of the volcano x-axis.}

\item{UpperLimitVolcano}{defaults to NULL. Allows to customize the upper limit of the volcano x-axis.}

\item{VolcanoLegendPosition}{defaults to "right". Allows to customize the position of the legend in the volcano plots.}

\item{FisherTermsDir}{directory where the file with the functional categories that will be used for the Fisher exact test of pathway enrichment are stored. The file must be a comma separated file (.csv). the first column contains the molecular entities and the subsequent columns the functional categories; each in a separate column, with the first row being the column name. For details see the exemplary file named xx.}

\item{returnSigFisher}{defaults to FALSE. When TRUE returns only the significant categories.}

\item{returnQvaluesFisher}{defaults to FALSE. When TRUE significances are only taken from adjusted P values, i.e., Q values.}

\item{h1Fisher}{defaults to "two.sided". Can be also "greater" or "less" and refers to whether functional categories are judged to be underrepresented and over-represented ("two.sided") as compared to all the significances, or whether only under-representation or over-representation is judged by the Fisher exact test.}

\item{ColorCategory}{name of the column in "FisherTermsDir" file from where the colors are taken for all the dots and categories in both plots.}

\item{PathwayColorsDir}{directory where the file defining colors for molecular entities and functional categories is located. The file must be tab separated (.txt). The first column contains the names of the molecular entities, the second column the functional category that will be used to implement colors and the third column R-hexadecimal color identifiers. For details see the exemplary file named xx.}

\item{BarNumbersPlot}{defaults to "pSigEntInClassToTotEntInClass" (proportion of significant entities in class as compared to total entities in class). Refers to the magnitude displayed in in the Fisher bar plots that show the pathway enrichment analysis results. Can also be "pEntInClassToTotEnt" "SigEntInCat".}

\item{FisherCatNumSize}{defaults to 3.5 points.}

\item{FisherCatAxisFontSize}{defaults to 10 points.}

\item{FisherAxesTitleSize}{defaults to 10 points.}

\item{FisherLegendFontSize}{defaults to 10 points.}

\item{FisherLegendTitleSize}{defaults to 10 points.}

\item{FisherLegendKeySize}{defaults to 10 points.}

\item{LowerLimitFisher}{defaults to NULL. Allows to customize the lower limit of the volcano x-axis.}

\item{UpperLimitFisher}{defaults to NULL. Allows to customize the lower limit of the volcano x-axis.}

\item{FisherLegendPosition}{defaults to "right". Allows to customize the position of the legend in the Fisher plots.}

\item{returnObject}{defaults to "VolcanoPlots", which returns a list with the ggplot objects. Can also be "FisherPlots" or "matrices".}
}
\description{
This function allows KineticMSI users to to summarize KineticMSI output in two complementary steps. First the function offers and option to draw volcano plots using the different previously calculated metrics as axes. For instance, P values, Q values, Tukey HSD Padj, Kolmogorov-Smirnov P values in the y-axis, and Log fold changes or Cohen's D statistic in the x-axis. from parameter two until 25, the function call allows to fine tune the volcano plot that comes out from the calculations, extending the customization for enhanced biological understanding. Second, the function allows to perform a pathway enrichment analyses given the appropriate file format with the pathway information (see exemplary file in ...). Parameters 26 to 41 allow to fine tune the details of the pathway enrichment analysis output, which is performed thorugh a Fisher exact test. The test inherits the full categories from the input files and hence supports its legitimacy on the biological accuracy of the files that are provided by the user. Finally, the function is able to return a list of plots with the desired outcome as well as the Fisher enriched categories, if present, in a .csv file named "FisherResults.csv".
}
\examples{

...
}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Results}
\keyword{Subsets}
\keyword{Summary}
\keyword{Tracer}
