% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kUserAssistedPartitionMSI.R
\name{kUserAssistedPartitionMSI}
\alias{kUserAssistedPartitionMSI}
\title{A function to manually subset consolidated data matrices into alike pixel groups using their distribution}
\usage{
kUserAssistedPartitionMSI(
  kAssesmentOutput,
  indexVector,
  ZeroAction = c(NULL, "remove", "replace")
)
}
\arguments{
\item{kAssesmentOutput}{output from the previous KineticMSI function. Namely, kAssessmentMSI.R. The object must be a list of matrices, one matrix for each molecular feature of interest measured across replicates and treatments (i.e., rows in each matrix). kAssesmentOutput must come from "minDataset" in the parameters of the call.}

\item{indexVector}{defines the index of molecular features inside the kAssesment object that will be manually partitioned.}

\item{ZeroAction}{allows to define what happens with zeros within the entity data matrices. If null, the zeros are fully preserved, if "remove" the rows that only contain zeros are deleted and if "replace" all zeros are replaced by normally distributed randomly generated numbers in the scale of 10^-12 to 10^-13.}
}
\description{
This function is interactive and needs to be run directly from the console to enhance the interactive mode. Thee function provides manual partition based on data distribution for molecular features that failed to be successfully subset using pvclust. The function will build a distribution density plot for each indexed molecular feature. The user then will be able to select how many partitions each feature has according to the histogram. Typically multi-modal distributions feature multiple peaks that can be captured this way. Subsequently the user will define and x-lim before which all pixels will be grabbed to integrate them into a subset; this procedure will be iterated for how many partitions the user previously defined. Finally, the function will return the mean partitions, just as the class discovery function before it.
}
\examples{

...
}
\keyword{Class}
\keyword{Discovery}
\keyword{Dynamics}
\keyword{MSI}
\keyword{Replicates}
\keyword{Tracer}
