% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIAcorrection.R
\name{NIAcorrection}
\alias{NIAcorrection}
\title{A function to correct natural isotopic abundances (NIA) from kineticMSI datasets or multiple csv files with the right format}
\usage{
NIAcorrection(
  MeasurementFileDir,
  pattern = "_rm0",
  SubSetReps = F,
  ElementFileDir,
  MoleculeFileDir,
  kCorrectTracerImpurity = TRUE,
  kCorrectTracerElementCore = TRUE,
  kCalculateMeanEnrichment = TRUE,
  kCorrectAlsoMonoisotopic = TRUE,
  kUltraHighRes = FALSE,
  kCalculationThreshold = 10^-8,
  kCalculationThreshold_UHR = 8,
  verbose = FALSE,
  outdir
)
}
\arguments{
\item{MeasurementFileDir}{directory where the input files are stored.}

\item{pattern}{defaults to "_rm0", that is, it grabs the files produced using rmNullPixel. It is used to define the pattern on which input files are looked for in the MeasurementFileDir}

\item{SubSetReps}{defaults to FALSE. Allows to subset the file list found in MeasurementFileDir.}

\item{ElementFileDir}{directory where the element input file is stored (see IsoCorrectoR documentation for input requirements).}

\item{MoleculeFileDir}{directory where the molecule input file is stored (see IsoCorrectoR documentation for input requirements).}

\item{kCorrectTracerImpurity}{defaults to TRUE.see IsoCorrectoR documentation for usage details.}

\item{kCorrectTracerElementCore}{defaults to TRUE. see IsoCorrectoR documentation for usage details.}

\item{kCalculateMeanEnrichment}{defaults to TRUE. see IsoCorrectoR documentation for usage details.}

\item{kCorrectAlsoMonoisotopic}{defaults to TRUE. see IsoCorrectoR documentation for usage details.}

\item{kUltraHighRes}{defaults to FALSE. see IsoCorrectoR documentation for usage details.}

\item{kCalculationThreshold}{defaults to IsoCorrectoR predefined value. see IsoCorrectoR documentation for usage details.}

\item{kCalculationThreshold_UHR}{defaults to IsoCorrectoR predefined value. see IsoCorrectoR documentation for usage details.}

\item{verbose}{defaults to FALSE. When TRUE it returns to the console the progression across the input files. Thus the parameter is meant to allow users to spot errors in the input files.}

\item{outdir}{defines the new directory that will contain IsoCorrectoR outputs, if the direction does not exist it will be automatically created. The new directory contains subdirectories named with the exact time of the run according to IsoCorrectoR conventions.}
}
\description{
This function allows you to correct isotopologue envelopes for NIA inheriting all specifications from the R package \href{https://www.bioconductor.org/packages/release/bioc/html/IsoCorrectoR.html}{IsoCorrectoR} The function calculates the percentage of enrichment of a defined stable isotope as well as other important values that reflect tracer dynamics within enrichment experiments. The function takes an entire directory and grabs all the csv files contained within in a recursive manner. Subsequently, the function generates csv files with each of the relevant returns in the \emph{kinetic}MSI context. Each column in the input table "MeasurementFile.csv" belongs to a single coordinate on the original image where the isotopologues could be measured and mined out.
}
\examples{

...
}
\keyword{IsoCorrectoR}
\keyword{KineticMSI}
\keyword{NIA}
\keyword{correction}
