% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProduceNewFiles.R
\name{ProduceNewFiles}
\alias{ProduceNewFiles}
\title{A function to produce new input files at multiple kineticMSI steps if the molecular species differ among the provided input files}
\usage{
ProduceNewFiles(reps)
}
\arguments{
\item{reps}{must be a list of replicate files.}
}
\description{
This function allows KineticMSI functions to produce new input files given uneven number of features.
}
\examples{

...
}
\keyword{KineticMSI}
\keyword{Preprocessing}
\keyword{troubleshooting}
