% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxySelection.R
\name{ProxySelection}
\alias{ProxySelection}
\title{A function to select an isotope proxy that best reflects the changes seen in steady state pools measured without stable isotope assisted mass spectrometry}
\usage{
ProxySelection(
  LabelledFileDir,
  TreatmentFileDir,
  NLSteadyStatePoolsDir,
  BatchCorr = TRUE,
  TreatmentIntoMeans = TRUE,
  Factor,
  Duplicate = FALSE
)
}
\arguments{
\item{LabelledFileDir}{directory where the enrichment-calculated feature steady state pool input file is stored.}

\item{TreatmentFileDir}{directory where a treatments csv file defining rows in the both labelled and non-labelled steady state pool files is located. Must contain a "Batch" column defining the batches to be corrected across files.}

\item{NLSteadyStatePoolsDir}{directory where the non-labelled feature steady state pool input file is stored.}

\item{BatchCorr}{defaults to TRUE. When TRUE, batch correction is performed.}

\item{TreatmentIntoMeans}{defaults to TRUE. When true, replicates are turned into means for the final diagnostic heatmap.}

\item{Factor}{factor vector that defines the treatment of each row in the steady state input files. Follows the same order as the treatments file.}

\item{Duplicate}{defaults to FALSE. Allows to correct batches of two replicates by duplicating the entries to enhance batch testing.}
}
\description{
This function allows to compare the steady state pools from molecular features that were measured through both stable assisted and conventional mass spectrometry. By contrasting the results from different isotope combinations (e.g., a0 to an versus a0 to a1) users can easily define how many isotopes reflect the actual pool changes of the target molecular features in their biological systems. The input files must have the same treatment rows while molecular feature may vary. The algorithm will select the features in common for the calculations.
}
\examples{

...
}
\keyword{Incorporation}
\keyword{Isotope}
\keyword{Proxy}
\keyword{Selection}
