% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNullPixel.R
\name{rmNullPixel}
\alias{rmNullPixel}
\title{A function to crop input kMSI datasets}
\usage{
rmNullPixel(
  MeasurementFileDir,
  pattern = "csv",
  SubSetReps = F,
  csvReturn = T,
  OnlyDeletePixelsWOIsotopologs = F,
  verbose = F,
  verboseFeature = F,
  rmDataStore = c("NewDir", "InputDir"),
  outdir = "rmOutput"
)
}
\arguments{
\item{MeasurementFileDir}{directory where the input files are stored.}

\item{pattern}{defaults to "csv". It is used to define the pattern on which input files are looked for in the MeasurementFileDir}

\item{SubSetReps}{defaults to FALSE. Allows to subset the file list found in MeasurementFileDir.}

\item{csvReturn}{defaults to TRUE. Returns corrected csv files.}

\item{OnlyDeletePixelsWOIsotopologs}{defaults to FALSE. When TRUE, it allows to only remove MSI pixels that lack isotopologues, while preserving those pixels that only have isotopologues, which  may be an indication of full stable isotope incorporation.}

\item{verbose}{defaults to FALSE. When TRUE it returns to the console the progression across the input files. Thus the parameter is meant to allow users to spot errors in the input files.}

\item{verboseFeature}{defaults to FALSE. When TRUE it returns to the console the progression across molecular variables. Thus the parameter is meant to allow users to spot errors in the input files.}

\item{rmDataStore}{needs to be defined. Specifies the output directory. Either a new directory "NewDir" or the same directory as the input "InputDir".}

\item{outdir}{only useful when rmDataStore = "NewDir". Defines the name of the new directory in which output is stored.}
}
\description{
This function allows you to remove MSI pixels that would impair interpretation of true 0\% enrichment in the downstream calculations. The function generates corrected csv files and a list with the corrected matrices as a return object in the R environment.The function takes an entire directory and it grabs all csv files within the provided directory. The function grabs each isotopologue envelope and sets to NA all of those pixels that would produce a misinterpretation of the NIA correction leading to misinterpreted enrichment percentages.
}
\examples{

...
}
\keyword{KineticMSI}
\keyword{Preprocessing}
