% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubSetInputFiles.R
\name{SubSetInputFiles}
\alias{SubSetInputFiles}
\title{A function to subset input files}
\usage{
SubSetInputFiles(reps, SubSetReps = T)
}
\arguments{
\item{reps}{must be a list of replicate files.}

\item{SubSetReps}{defaults to TRUE. Allows to subset the file list found in MeasurementFileDir.}
}
\description{
This function allows KineticMSI functions to subset lists of files at multiple steps.
}
\examples{

...
}
\keyword{KineticMSI}
\keyword{Preprocessing}
